requirements.txt
fastapi
streamlit
accelerate
uvicorn
python-multipart
Pillow
librosa
soundfile
transformers
datasets
pydub
yt_dlp
pyannote.audio
pyannote.database
pyannote.metrics
pyannote.core
lightning
python-dotenv
optimum